# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order API",
        "description": "API для заказов",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8889",
            "description": "Local API server"
        }
    ],
    "paths": {
        "/api/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Создание нового заказа",
                "operationId": "2123ba58dadb22bd6f2c8081af562230",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_name",
                                    "customer_email",
                                    "products"
                                ],
                                "properties": {
                                    "customer_name": {
                                        "type": "string",
                                        "example": "Иван Иванов"
                                    },
                                    "customer_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ivan@example.com"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "minimum": 1,
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешное создание заказа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Получить заказ по ID",
                "operationId": "88d27387b1e409e075c18609d4a0c300",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заказа",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о заказе",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заказ не найден"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "OrderResource": {
                "title": "Order Response",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "customer_name": {
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "customer_email": {
                        "type": "string",
                        "example": "ivan@example.com"
                    },
                    "total_price": {
                        "type": "number",
                        "format": "float",
                        "example": 1234.56
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "product_id": {
                                    "type": "integer",
                                    "example": 3
                                },
                                "product_name": {
                                    "type": "string",
                                    "example": "Product 3"
                                },
                                "quantity": {
                                    "type": "integer",
                                    "example": 2
                                },
                                "unit_price": {
                                    "type": "number",
                                    "format": "float",
                                    "example": 120
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Orders",
            "description": "Orders"
        }
    ]
}